backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
pydantic-settings>=2.0.0
youtube-transcript-api>=1.0.0
yt-dlp>=2025.1.0
openai>=1.40.0
anthropic>=0.34.0
celery>=5.4.0
redis>=5.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
supabase>=2.5.0
PyJWT>=2.8.0
slowapi>=0.1.9
sentry-sdk[fastapi]>=2.27.0
fpdf2>=2.7.0
markdown>=3.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
black
ruff
